+++ pom.xml
@@ -16,7 +16,7 @@
         <aspectj.version>1.9.7</aspectj.version>
         <allure.maven>2.11.2</allure.maven>
         <jdi.version>RELEASE</jdi.version>
-        <log4j2.version>2.17.1</log4j2.version>
+        <log4j2.version>2.19.0</log4j2.version>
         <remote.url/>
         <!-- - -add-opens java.base/java.lang=ALL-UNNAMED needed for java 16+-->
         <aspectj.args></aspectj.args>
@@ -27,6 +27,12 @@
             <id>ci</id>
             <properties>
                 <!--<remote.url>http://localhost:4444/wd/hub</remote.url>-->
+            </properties>
+        </profile>
+        <profile>
+            <id>jdk16</id>
+            <properties>
+                <aspectj.args>--add-opens java.base/java.lang=ALL-UNNAMED</aspectj.args>
             </properties>
         </profile>
     </profiles>
@@ -40,9 +46,19 @@
         </dependency>
 
         <dependency>
+            <groupId>io.cucumber</groupId>
+            <artifactId>cucumber-junit</artifactId>
+            <version>7.10.1</version>
+        </dependency>
+        <dependency>
+            <groupId>io.cucumber</groupId>
+            <artifactId>cucumber-core</artifactId>
+            <version>7.10.1</version>
+        </dependency>
+        <dependency>
             <groupId>io.qameta.allure</groupId>
-            <artifactId>allure-cucumber-jvm</artifactId>
-            <version>2.19.0</version>
+            <artifactId>allure-cucumber7-jvm</artifactId>
+            <version>2.20.1</version>
         </dependency>
         <dependency>
             <groupId>io.qameta.allure</groupId>
@@ -70,23 +86,6 @@
             <artifactId>log4j-1.2-api</artifactId>
             <version>${log4j2.version}</version>
         </dependency>
-
-        <dependency>
-            <groupId>io.cucumber</groupId>
-            <artifactId>cucumber-junit</artifactId>
-            <version>7.10.1</version>
-        </dependency>
-        <dependency>
-            <groupId>io.cucumber</groupId>
-            <artifactId>cucumber-core</artifactId>
-            <version>7.10.1</version>
-        </dependency>
-        <dependency>
-            <groupId>io.qameta.allure</groupId>
-            <artifactId>allure-cucumber7-jvm</artifactId>
-            <version>2.20.1</version>
-        </dependency>
-
     </dependencies>
 
     <build>
